+++ services/backend/src/modules/symptoms/symptoms.service.ts
@@ -7,26 +7,42 @@ import { UpdateSymptomDto } from './dto/update-symptom.dto';
 export class SymptomsService {
   constructor(private prisma: PrismaService) {}
 
-  async create(createSymptomDto: CreateSymptomDto) {
-    return this.prisma.symptom.create({ data: createSymptomDto });
-  }
-
-  async findAll() {
-    return this.prisma.symptom.findMany();
-  }
-
-  async findById(uuid: string) {
-    return this.prisma.symptom.findUnique({ where: { uuid } });
-  }
-
-  async update(uuid: string, updateSymptomDto: UpdateSymptomDto) {
-    return this.prisma.symptom.update({
-      where: { uuid },
-      data: updateSymptomDto,
+  async create(createSymptomDto: CreateSymptomDto, userId: string) {
+    return this.prisma.symptom.create({
+      data: {
+        ...createSymptomDto,
+        user: {
+          connect: { uuid: userId }, // Connect the User using the provided userId
+        },
+      },
     });
   }
 
-  async delete(uuid: string) {
-    return this.prisma.symptom.delete({ where: { uuid } });
+  async findAll(userId: string) {
+    return this.prisma.symptom.findMany({ where: { userId } });
+  }
+
+  async findById(uuid: string, userId: string) {
+    return this.prisma.symptom.findUnique({ where: { uuid, userId } });
+  }
+
+  async update(
+    uuid: string,
+    updateSymptomDto: UpdateSymptomDto,
+    userId: string,
+  ) {
+    return this.prisma.symptom.update({
+      where: { uuid },
+      data: {
+        ...updateSymptomDto,
+        user: {
+          connect: { uuid: userId }, // Connect the User using the provided userId
+        },
+      },
+    });
+  }
+
+  async delete(uuid: string, userId: string) {
+    return this.prisma.symptom.delete({ where: { uuid, userId } });
   }
 }
